# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Scan -\u003e Plan -\u003e Approve -\u003e Fix -\u003e Rollback REST API for golangci-lint issue triage.",
        "title": "Golangci Lint Dashboard API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/issues/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Issues"
                ],
                "summary": "Get a lint issue by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Issue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.LintIssue"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
...
        "/scans": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Enqueues a golangci-lint scan for repo_ref+branch. Rule BR-1: only one scan or fix may run at a time per repo+branch.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scans"
                ],
                "summary": "Start a lint scan",
                "parameters": [
                    {
                        "description": "Scan target",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createScanRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "409": {
                        "description": "a scan or fix is already running for this repo+branch",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/scans/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scans"
                ],
                "summary": "Get a scan by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.LintScan"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/scans/{id}/issues": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scans"
                ],
                "summary": "List issues found by a scan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 50, capped at 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.LintIssue"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.createScanRequest": {
            "type": "object",
            "required": [
                "branch",
                "repo_ref"
            ],
            "properties": {
                "branch": {
                    "type": "string"
                },
                "repo_ref": {
                    "type": "string"
                }
            }
        },
        "golangci_backend_frontiir_utils.RestErr": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable business error code (e.g. \"SLA_EXPIRED\")",
                    "type": "string"
                },
                "data": {
                    "description": "Optional structured payload (e.g. conflicting ticket detail)"
                },
                "message": {
                    "description": "User-friendly error message",
                    "type": "string"
                },
                "status": {
                    "description": "HTTP status code",
                    "type": "integer"
                }
            }
        },
...
        "storage.FixPlan": {
            "type": "object",
            "properties": {
                "acceptance_criteria": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "after_snippet": {
                    "type": "string"
                },
                "approved_at": {
                    "type": "string"
                },
                "approved_by": {
                    "description": "ApprovedBy/ApprovedAt implement Rule BR-3's \"approver identity +\ntimestamp recorded\" requirement (10-business-rules.md) -- absent\nfrom the original M1 schema. ApprovedAt is nil until approve/reject.",
                    "type": "string"
                },
                "batch_fingerprint": {
                    "description": "BatchFingerprint/PromptVersion implement the Risk register's own\nmitigation (\"cache plan by issue+source fingerprint; store model +\nprompt version on FixPlan\") -- absent from the original M1 schema.",
                    "type": "string"
                },
                "before_snippet": {
                    "type": "string"
                },
                "breaking_change": {
                    "type": "boolean"
                },
                "code_context": {
                    "description": "CodeContext/FixStrategyCode/BeforeSnippet/AfterSnippet/SideEffects/\nImpactAnalysis/RecommendedTestCommands/AcceptanceCriteria added\n2026-08-19 to match before-fixed/*.md's section depth (§2, §4-9) --\nsee planner.PlanResult's matching doc comments for what populates each.",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_behavior": {
                    "description": "required by 04-prd.md FR-4 / 05-ui-design.md's Plan Viewer -- missing from the original M1 schema",
                    "type": "string"
                },
                "current_behavior_my": {
                    "type": "string"
                },
                "files_impacted": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "fix_strategy_code": {
                    "type": "string"
                },
                "generated_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "impact_analysis": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "issues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.LintIssue"
                    }
                },
                "prompt_version": {
                    "type": "string"
                },
                "recommended_fix": {
                    "type": "string"
                },
                "recommended_fix_my": {
                    "type": "string"
                },
                "recommended_test_commands": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "risk_level": {
                    "description": "low|medium|high",
                    "type": "string"
                },
                "root_cause": {
                    "type": "string"
                },
                "root_cause_my": {
                    "description": "RootCauseMy/CurrentBehaviorMy/RecommendedFixMy are Burmese counterparts\nto the 3 fields above, for the UI's language toggle -- additive, the\nEnglish fields are unchanged and remain the fallback when these are empty.",
                    "type": "string"
                },
                "side_effects": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "description": "generating|pending|approved|rejected|applied|failed -- \"generating\"\nand \"failed\" added in M3 for the async plan-job sub-states (a client\npolling GET /plans/:id needs to tell \"still working\" from\n\"permanently failed, e.g. AI Layer unavailable\" apart); the other 4\nare 08-validation-rules.md's originally decided enum.",
                    "type": "string"
                },
                "test_plan": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "storage.LintIssue": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "fingerprint": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "line": {
                    "type": "integer"
                },
                "linter": {
                    "type": "string"
                },
                "message": {
                    "description": "golangci-lint's issue text, minus the rule prefix -- see parser.Parse",
                    "type": "string"
                },
                "plans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.FixPlan"
                    }
                },
                "reason_my": {
                    "description": "Burmese \"why this breaks the rule\" explanation -- see parser.reasonMy",
                    "type": "string"
                },
                "rule": {
                    "type": "string"
                },
                "scan_id": {
                    "type": "string"
                },
                "severity": {
                    "description": "critical|high|medium|low|info",
                    "type": "string"
                },
                "status": {
                    "description": "open|planned|fix_applied|resolved|reopened|ignored",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "storage.LintScan": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "commit_sha": {
                    "type": "string"
                },
                "config_hash": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "raw_output_ref": {
                    "type": "string"
                },
                "repo_ref": {
                    "type": "string"
                },
                "status": {
                    "description": "running|success|failed",
                    "type": "string"
                },
                "triggered_by": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}